feat: moved MusdConversionCta into its own directory
